refactor(plugins): separate runtime bindings from artifact selection
Keep Doctor filesystem selection independent of runtime record types, migrate every binding caller, and verify repaired persistence through the built CLI instead of a source-mode observer.
